requirements.txt
# Core
django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-filter>=23.5

# Database
psycopg2-binary>=2.9

# Celery & Redis
celery>=5.3
redis>=5.0
django-celery-beat>=2.6
django-celery-results>=2.5

# API Docs
drf-spectacular>=0.27

# Utilities
Pillow>=10.0
python-decouple>=3.8
django-extensions>=3.2

# Dev / Testing
pytest>=7.4
pytest-django>=4.7
factory-boy>=3.3
